requirements.txt
python-telegram-bot==22.0
apscheduler==3.10.4
aiosqlite==0.20.0
aiohttp==3.9.5
requests==2.31.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
colorlog==6.8.2
pytrends==4.9.2
fredapi==0.5.2
